fix old spectator view
