+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/SpectatorViewOld.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25b080c0fe9f22449a6923148a1afef2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
